pyproject.toml
[project]
name = "rfp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.25.2",
    "fastapi>=0.133.0",
    "github-copilot-sdk>=0.1.25",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
